Move `checked_load` utilities to separate file to avoid circular imports
